Increased postDt, tstop and total radius to see how much it propagates in a larger fuel mass.
It definatly ignights as a burn wave travels through the fuel but it looses the shock wave so maybe the critical burn up parameter is to do with that?

The shock wave is associated with detonation but the thermal wave is faster at higher temperatures.

As TN burning is established it could be driven by alpha heating too